requirements.txt
# Core
CherryPy==3.2.6
sqlalchemy==2.0.45
psycopg2-binary>=2.9.9
python-ldap==3.1.0
pyOpenSSL==19.1.0

# Serialization / Utilities
IPy==1.01
Cheetah3>=3.2.7b1

# Optional or built-in replacements
# python-processing  # built-in for Python >=2.6
# pydhcplib is included in source tree
